# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReliefOps API — crisis logistics for NGOs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List all supply requests
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplyRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a supply request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyRequest"
  /requests/{requestId}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get request detail
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Request with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyRequestDetail"
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update request status/priority
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestBody"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyRequest"
  /requests/{requestId}/matching-hubs:
    get:
      operationId: getMatchingHubs
      tags: [requests]
      summary: Get candidate hubs ranked by straight-line distance for a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ranked hub matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HubMatch"

  # ─── Transfers ───────────────────────────────────────────────────────────────
  /transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      summary: List all transfers
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: requestId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      operationId: createTransfer
      tags: [transfers]
      summary: Create a transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferBody"
      responses:
        "201":
          description: Created transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{transferId}:
    get:
      operationId: getTransfer
      tags: [transfers]
      summary: Get transfer detail
      parameters:
        - name: transferId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferDetail"
    patch:
      operationId: updateTransfer
      tags: [transfers]
      summary: Update transfer status (triggers stock reduction on Dispatched)
      parameters:
        - name: transferId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransferBody"
      responses:
        "200":
          description: Updated transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"

  # ─── Volunteers ──────────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks with optional filters
      parameters:
        - name: volunteerId
          in: query
          required: false
          schema:
            type: string
        - name: transferId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create and assign a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task status or assignment
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  # ─── Activity Log ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Hub:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
          nullable: true
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
...
    Item:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [Medicine, Food, Hygiene, "First Aid"]
        barcode:
          type: string
          nullable: true
        unit:
          type: string
        imageUrl:
          type: string
          nullable: true
        tracksExpiry:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - unit
        - tracksExpiry
        - createdAt
...
    SupplyRequest:
      type: object
      properties:
        id:
          type: string
        requestingHubId:
          type: string
        requestingHub:
          $ref: "#/components/schemas/Hub"
        priority:
          type: string
          enum: [Low, Medium, Urgent, Critical]
        status:
          type: string
          enum: [Draft, Open, Assigned, Dispatched, Delivered]
        notes:
          type: string
          nullable: true
        createdBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - requestingHubId
        - requestingHub
        - priority
        - status
        - createdAt
    SupplyRequestDetail:
      allOf:
        - $ref: "#/components/schemas/SupplyRequest"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/RequestItem"
            transfers:
              type: array
              items:
                $ref: "#/components/schemas/Transfer"
          required:
            - items
            - transfers
    RequestItem:
      type: object
      properties:
        id:
          type: string
        requestId:
          type: string
        itemId:
          type: string
        item:
          $ref: "#/components/schemas/Item"
        quantityNeeded:
          type: integer
      required:
        - id
        - requestId
        - itemId
        - item
        - quantityNeeded
    CreateRequestBody:
      type: object
      properties:
        requestingHubId:
          type: string
        priority:
          type: string
          enum: [Low, Medium, Urgent, Critical]
        status:
          type: string
          enum: [Draft, Open, Assigned, Dispatched, Delivered]
        notes:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: string
              quantityNeeded:
                type: integer
            required:
              - itemId
              - quantityNeeded
      required:
        - requestingHubId
        - priority
    UpdateRequestBody:
      type: object
      properties:
        priority:
          type: string
          enum: [Low, Medium, Urgent, Critical]
        status:
          type: string
          enum: [Draft, Open, Assigned, Dispatched, Delivered]
        notes:
          type: string
    HubMatch:
      type: object
      properties:
        hub:
          $ref: "#/components/schemas/Hub"
        distanceKm:
          type: number
        availableItems:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: string
              itemName:
                type: string
              availableQuantity:
                type: integer
              requestedQuantity:
                type: integer
            required:
              - itemId
              - itemName
              - availableQuantity
              - requestedQuantity
        canFulfillAll:
          type: boolean
      required:
        - hub
        - distanceKm
        - availableItems
        - canFulfillAll

    # ─── Transfer ─────────────────────────────────────────────────────────────
    Transfer:
      type: object
      properties:
        id:
          type: string
        requestId:
          type: string
        fromHubId:
          type: string
        fromHub:
          $ref: "#/components/schemas/Hub"
        toHubId:
          type: string
        toHub:
          $ref: "#/components/schemas/Hub"
        status:
          type: string
          enum: [Planned, Dispatched, Delivered]
        etaMinutes:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - requestId
        - fromHubId
        - fromHub
        - toHubId
        - toHub
        - status
        - createdAt
    TransferDetail:
      allOf:
        - $ref: "#/components/schemas/Transfer"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/TransferItem"
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
          required:
            - items
            - tasks
    TransferItem:
      type: object
      properties:
        id:
          type: string
        transferId:
          type: string
        itemId:
          type: string
        item:
          $ref: "#/components/schemas/Item"
        quantity:
          type: integer
      required:
        - id
        - transferId
        - itemId
        - item
        - quantity
    CreateTransferBody:
      type: object
      properties:
        requestId:
          type: string
        fromHubId:
          type: string
        toHubId:
          type: string
        etaMinutes:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: string
              quantity:
                type: integer
            required:
              - itemId
              - quantity
      required:
        - requestId
        - fromHubId
        - toHubId
        - items
    UpdateTransferBody:
      type: object
      properties:
        status:
          type: string
          enum: [Planned, Dispatched, Delivered]
        etaMinutes:
          type: integer

    # ─── Volunteer ────────────────────────────────────────────────────────────
    Volunteer:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        hasVehicle:
          type: boolean
        availabilityStatus:
          type: string
          enum: [Available, Busy, Offline]
        activeTasks:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - hasVehicle
        - availabilityStatus
        - activeTasks
        - createdAt
...
    Task:
      type: object
      properties:
        id:
          type: string
        transferId:
          type: string
          nullable: true
        volunteerId:
          type: string
          nullable: true
        volunteer:
          $ref: "#/components/schemas/Volunteer"
          nullable: true
        type:
          type: string
          enum: [Pickup, "Delivery/Transfer", "Stock Count"]
        status:
          type: string
          enum: [Open, Assigned, "In Progress", Done]
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - status
        - createdAt
    CreateTaskBody:
      type: object
      properties:
        transferId:
          type: string
        volunteerId:
          type: string
        type:
          type: string
          enum: [Pickup, "Delivery/Transfer", "Stock Count"]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
      required:
        - type
    UpdateTaskBody:
      type: object
      properties:
        volunteerId:
          type: string
        status:
          type: string
          enum: [Open, Assigned, "In Progress", Done]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time

    # ─── Activity ─────────────────────────────────────────────────────────────
...